fix(cli): move api-key check to hatch, use provider catalog, fix Linux PATH
- Move the LLM provider API key warning from `wake` to `hatchLocal` so
  new users are told about a missing key at first-time setup, not on
  every subsequent restart (per maintainer review feedback)
- Use `LLM_PROVIDER_ENV_VAR_NAMES` from the canonical provider env-var
  catalog so the checked keys stay in sync automatically as new
  providers are added (covers anthropic, openai, gemini, fireworks,
  openrouter)
- Remove the check and import from `wake.ts` entirely
- Update `setup.sh` to append `~/.bun/bin` to `~/.bashrc` / `~/.zshrc`
  if not already present, eliminating the "vellum not found" pain point
  for Linux contributors without extra manual steps
- Simplify README Linux notes: replace the three-item list with two
  items, add `source ~/.bashrc` to the install-from-source snippet, and
  drop the stale section about `assistant/.env` (hatch now prompts for
  the key directly)
- Update tests to match the new provider set (swap OLLAMA for FIREWORKS
  and OPENROUTER)
